+++ app/src/main/java/com/pp/masterand/data/Repositories.kt
@@ -4,7 +4,7 @@ import kotlinx.coroutines.flow.Flow
 
 interface PlayersRepository {
     fun getAllPlayersStream(): Flow<List<Player>>
-    fun getPlayerStream(id: Int): Flow<Player?>
+    fun getPlayerStream(id: Long): Flow<Player?>
     suspend fun getPlayersByEmail(email: String): List<Player>
     suspend fun insertPlayer(player: Player) : Long
 }
@@ -13,7 +13,7 @@ class PlayersRepositoryImpl(private val playerDao: PlayerDao) : PlayersRepositor
         TODO("Not yet implemented")
     }
 
-    override fun getPlayerStream(playerId: Int): Flow<Player?> =
+    override fun getPlayerStream(playerId: Long): Flow<Player?> =
         playerDao.getPlayerStream(playerId)
     override suspend fun getPlayersByEmail(email: String): List<Player> =
         playerDao.getPlayersByEmail(email)
